% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ccp}
\alias{ccp}
\title{Remove NULL List Elements}
\usage{
ccp(x)
}
\arguments{
\item{x}{'list'.}
}
\value{
List
}
\description{
Removes elements from a list that are equal to \code{NULL}.
}
\examples{
d <- list("a", "b", NULL, "c")
d
d <- ccp(d)
d
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
