% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_package}
\alias{check_package}
\title{Check Package}
\usage{
check_package(pkg, msg = NULL, call. = FALSE)
}
\arguments{
\item{pkg}{'character' string.
Package name.}

\item{msg}{'character' string.
Action package is used for.}

\item{call.}{'logical' flag.
Whether the function should be part of the error message.}
}
\value{
Stops execution if the package is missing from the name space.
}
\description{
Check that a package is available.
}
\examples{
check_package(pkg = "inldata", msg = "Dataset access")
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
