% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tables.R
\name{combine_tables}
\alias{combine_tables}
\title{Combine Package Datasets}
\usage{
combine_tables(package)
}
\arguments{
\item{package}{'character' string.
Package name.}
}
\value{
A list of dataset elements of class 'data.frame'.
A list element is named using its dataset name.
}
\description{
Combine package datasets that can be coerced into a data frame into a single list.
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
