% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download File from the Internet}
\usage{
download_file(url, cachedir = get_cache_dir(), ..., quiet = FALSE, mode = "wb")
}
\arguments{
\item{url}{'character' string.
URL of a resource to be downloaded.}

\item{cachedir}{'character' string.
Path to the cache directory, see \code{\link{get_cache_dir}} function for the default value.
The file will not be downloaded if the file exists in the cache directory.}

\item{...}{Additioanl arguments to be passed to the \code{\link[utils:download.file]{download.file}} function.
The destination of the downloaded file may not be specified.}

\item{quiet}{'logical' flag.
Whether to supress status messages (if any), and the progress bar.}

\item{mode}{'character' string.
Mode with which to write the file. Useful values are "w", "wb" (binary), "a" (append) and "ab".}
}
\value{
Returns the path to the downloaded file, or extracted files when decompression occurs.
}
\description{
Download a file from the Internet and save to a cache directory.
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
