% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data_list.R
\name{filter_data_list}
\alias{filter_data_list}
\title{Filter Data List Column}
\usage{
filter_data_list(table_id, style = "width: 100\%; height: 28px;")
}
\arguments{
\item{table_id}{'character' string.
Unique table identifier.}

\item{style}{'character' string.
CSS style applied to input HTML tag.}
}
\value{
Returns a function to perform filtering.
}
\description{
Create a data list column filter for a React Table.
Requires that the \pkg{htmltools} packages is available.
}
\examples{
f <- filter_data_list("table-id")
}
\keyword{internal}
