% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_dir.R
\name{get_cache_dir}
\alias{get_cache_dir}
\alias{clear_cache_dir}
\title{Manage Cache Directory}
\usage{
get_cache_dir(name = "inldata")

clear_cache_dir(name = "inldata")
}
\arguments{
\item{name}{'character' string.
Name of cache directory.}
}
\value{
Path to the cache directory.
}
\description{
Find, create, or clear the cache directory.
Defaults to the temporary directory if the \pkg{rappdirs} package is unavailable.
You can specify the path to the cache directory by setting an environment variable named "CACHE_DIR".
}
\examples{
dir <- get_cache_dir("test")

clear_cache_dir("test")
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
