% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Extract File Extension}
\usage{
get_file_ext(x, compression = TRUE)
}
\arguments{
\item{x}{'character' vector.
File path(s).}

\item{compression}{'logical' flag.
Whether to account for the compression extension '.gz', '.bz2' or '.xz'.}
}
\value{
Returns the file (name) extensions (excluding the leading dot).
}
\description{
Extract the extension of a file.
}
\examples{
c("dir/file.txt", "file.txt.gz") |>
  get_file_ext()
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
