% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_size}
\alias{get_file_size}
\title{Extract File Size}
\usage{
get_file_size(paths)
}
\arguments{
\item{paths}{'character' vector.
File paths.}
}
\value{
Formatted file size(s).
}
\description{
Extract size on the user's file system.
}
\examples{
system.file("CITATION", package = "inldata") |>
  get_file_size()
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
