% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inl.R
\docType{data}
\name{inl}
\alias{inl}
\title{Idaho National Laboratory Boundary}
\format{
A SpatialPolygonsDataFrame of the \pkg{sp} package object with 1 feature and 4 variables.
  See \code{\link{projection}} dataset for coordinate reference system information.
}
\source{
U.S. Geological Survey
  \href{https://www.usgs.gov/centers/id-water/science/idaho-national-laboratory-project-office}{Idaho National Laboratory Project Office}
}
\usage{
inl
}
\description{
Boundary of the Idaho National Laboratory (\href{https://inl.gov/}{INL}).
}
\examples{
inlmisc::PlotMap(inl, dms.tick = TRUE)
sp::plot(inl, col = "red", add = TRUE)
str(inl@data)

}
\keyword{datasets}
