% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwd.R
\docType{data}
\name{iwd}
\alias{iwd}
\title{Industrial Waste Ditch}
\format{
A \href{https://r-spatial.github.io/sf/articles/sf1.html}{simple feature} with fields:
\describe{
\item{\code{geometry}}{Sequence of points connected by straight, non-self-intersecting line pieces,
one-dimensional geometry.}
}
}
\source{
A GeoJSON file of the industrial waste ditch was created by the U.S. Geological Survey
\href{https://www.usgs.gov/centers/idaho-water-science-center/science/idaho-national-laboratory-project-office}{Idaho National Laboratory Project Office}
}
\usage{
iwd
}
\description{
An industrial waste ditch located near the Naval Reactors Facility (NRF) in eastern Idaho.
The NRF discharges its wastewater to this 1.65-mile-long waste ditch.
The waste ditch was initially used in 1953 to dispose of non-radioactive, non-sewage industrial wastewater.
}
\examples{
print(iwd)
}
\keyword{datasets}
