% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakes.R
\docType{data}
\name{lakes}
\alias{lakes}
\title{Lakes and Ponds}
\format{
A SpatialPolygonsDataFrame of the \pkg{sp} package with 155 features of 12 variables.
  See \code{\link{projection}} dataset for coordinate reference system information.
}
\source{
U.S. Geological Survey (USGS), National Geospatial Technical Operations Center,
  USGS National Hydrography Dataset (NHD) Medium Resolution for Idaho,
  released August 4, 2014.
}
\usage{
lakes
}
\description{
Perennial lakes and ponds in the vicinity of Idaho National Laboratory, eastern Idaho.
}
\examples{
inlmisc::PlotMap(lakes, dms.tick = TRUE,
                 lakes = list(lakes, "lwd" = 1))
str(lakes@data)

}
\keyword{datasets}
