% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountains.R
\docType{data}
\name{mountains}
\alias{mountains}
\title{Mountain Ranges and Buttes}
\format{
A SpatialPolygonsDataFrame of the \pkg{sp} package with 17 features and 1 variable.
  See \code{\link{projection}} dataset for coordinate reference system information.
}
\source{
U.S. Geological Survey
  \href{https://www.usgs.gov/centers/id-water/science/idaho-national-laboratory-project-office}{Idaho National Laboratory Project Office}
}
\usage{
mountains
}
\description{
Simplified representation of the mountain ranges and buttes in the
vicinity of Idaho National Laboratory, eastern Idaho.
}
\examples{
inlmisc::PlotMap(mountains, dms.tick = TRUE)
sp::plot(mountains, col = "gray", add = TRUE)
str(mountains@data)

}
\keyword{datasets}
