% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rd_db.R
\name{parse_rd_db}
\alias{parse_rd_db}
\title{Parse R-Package Dataset Documentation}
\usage{
parse_rd_db(..., doc_type = "data")
}
\arguments{
\item{...}{Arguments to be passed to the \code{\link[tools:Rdutils]{tools::Rd_db}} function,
such as \code{package}, a character string naming an installed package.}

\item{doc_type}{'character' vector.
Document type, such as \code{data} for dataset objects.}
}
\value{
A named list containing the documentation sections as character strings.
}
\description{
Reads and parses R documentation (Rd) files.
}
\examples{
l <- parse_rd_db(package = "inldata")
str(l, 1)
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
