% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_station_nm}
\alias{parse_station_nm}
\title{Parse Station Names}
\usage{
parse_station_nm(x)
}
\arguments{
\item{x}{'character' vector.
Station names, such as \verb{03N 29E 01DBB1    USGS 98}.}
}
\value{
A vector of common site names parsed from \code{x}.
}
\description{
Convert station names to common site names.
}
\examples{
x <- c(
  "02N 30E 08DAD1               BADGING FACILITY",
  "03N 29E 33CCC1 MIDDLE 2051 PORT6 ZONE6 826.8FT",
  "MUD LAKE NR TERRETON ID"
)
parse_station_nm(x)
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
