% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_pkg_dir}
\alias{test_pkg_dir}
\title{Test Location is Package Directory}
\usage{
test_pkg_dir(pkg)
}
\arguments{
\item{pkg}{'character' string.
Package name.}
}
\value{
Formatted file size(s).
}
\description{
Test whether the working directory is located at the top level directory of the package.
}
\examples{
test_pkg_dir("inldata")
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
