% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_station_nm}
\alias{trim_station_nm}
\title{Trim Station Names}
\usage{
trim_station_nm(x)
}
\arguments{
\item{x}{'character' vector.
Station names, such as \verb{03N 29E 01DBB1    USGS 98}.}
}
\value{
A vector of common site names parsed from \code{x}.
}
\description{
Convert station names to common site names.
}
\examples{
x <- c(
  "03N 29E 12DDB1    FIRE STA 2",
  "03N 29E 24CCA1 MIDDLE 2050A PORT15 ZONE15 516.8FT"
)
trim_station_nm(x)
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
