% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\docType{data}
\name{units}
\alias{units}
\title{Parameter Units}
\format{
A data frame with columns:
\describe{
\item{\code{parm_unit}}{Parameter units abbreviation.}
\item{\code{desc}}{Parameter units description.}
\item{\code{siunitx}}{Parameter units formatted for LaTeX using the
\href{https://ctan.org/pkg/siunitx}{siunitx} package notation.}
}
}
\source{
Idaho National Laboratory Project Office
}
\usage{
units
}
\description{
Information describing the parameter units of measure.
}
\examples{
str(units)
}
\keyword{datasets}
