% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountains.R
\docType{data}
\name{mountains}
\alias{mountains}
\title{Mountain Ranges and Buttes}
\format{
A \href{https://r-spatial.github.io/sf/articles/sf1.html}{simple feature} with fields:
\describe{
\item{\code{name}}{Feature name.}
\item{\code{geometry}}{Polygon geometry with a positive area (two-dimensional);
sequence of points that form a closed, non-self-intersecting ring;
the first ring denotes the exterior ring,
zero or more subsequent rings denote holes in this exterior ring.}
}
}
\source{
Spatial polygons were created from areas in the digital elevation model (DEM)
where terrain slope is greater than a specified threshold.
}
\usage{
mountains
}
\description{
A simplified representation of mountain ranges and buttes in the
vicinity of Idaho National Laboratory (INL) in eastern Idaho.
}
\examples{
print(mountains)
}
\keyword{datasets}
