% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\docType{data}
\name{streams}
\alias{streams}
\title{Rivers and Streams}
\format{
A SpatialLinesDataFrame of the \pkg{sp} package with 197 features of 14 variables.
  See \code{\link{projection}} dataset for coordinate reference system information.
}
\source{
U.S. Geological Survey (USGS), National Geospatial Technical Operations Center,
  USGS National Hydrography Dataset (NHD) Medium Resolution for Idaho,
  released August 4, 2014.
}
\usage{
streams
}
\description{
Stream segments in the vicinity of Idaho National Laboratory, eastern Idaho.
}
\examples{
inlmisc::PlotMap(streams, dms.tick = TRUE,
                 rivers = list(streams, "lwd" = 1))
str(streams@data)

}
\keyword{datasets}
