% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\docType{data}
\name{units}
\alias{units}
\title{Units of Measurement}
\format{
A data frame with columns:
\describe{
\item{\code{unit_cd}}{Measurment abbreviation for units.}
\item{\code{unit_ds}}{Description for units of measurement.}
\item{\code{siunitx}}{Units formatted for LaTeX using the
\href{https://ctan.org/pkg/siunitx}{siunitx} package notation.}
}
}
\source{
Idaho National Laboratory Project Office
}
\usage{
units
}
\description{
Descriptive information about the measurement units.
}
\examples{
str(units)
}
\keyword{datasets}
